refactor: extract gas fee estimation logic to new component
